Working on a base version of the /api/entities endpoint
