+++ vanilla/apis/commons.py
@@ -1,0 +1,32 @@
+# -*- coding: utf-8 -*-
+
+"""
+Common functions for EUDAT endpoints
+"""
+
+from __future__ import absolute_import
+
+from ..base import ExtendedApiResource
+from commons.logs import get_logger
+
+logger = get_logger(__name__)
+
+
+class EudatEndpoint(ExtendedApiResource):
+
+    def init_endpoint(self):
+
+        #####################################
+        ## WHEN THE GRAPH WILL BE INTEGRATED
+        # # Note: graph holds the authenticated accounts in our architecture
+        # graph = self.global_get_service('neo4j')
+        # graphuser = self.get_current_user()
+        # irodsuser = icom.translate_graph_user(graph, graphuser)
+        # icom = self.global_get_service('irods', user=irodsuser.username)
+
+        #####################################
+# // TO FIX:
+# understand how to know which user is connected
+        icom = self.global_get_service('irods')
+
+        return icom
